+++ src/Ydb.Sdk/test/Ydb.Sdk.Topic.Tests/Utils.cs
@@ -6,4 +6,6 @@ public static class Utils
 {
     public static readonly ILoggerFactory LoggerFactory =
         Microsoft.Extensions.Logging.LoggerFactory.Create(builder => builder.AddConsole());
+
+    public const string ConnectionString = "Host=localhost;Port=2136;Database=/local";
 }